programação funcional

+++ modulo-20/src/entities/Product.java
@@ -1,12 +1,13 @@
 package entities;
 
+import java.math.BigDecimal;
 import java.util.Comparator;
 
 public class Product {
     private String name;
-    private Double price;
+    private BigDecimal price;
 
-    public Product(String name, Double price) {
+    public Product(String name, BigDecimal price) {
         this.name = name;
         this.price = price;
     }
@@ -19,11 +20,11 @@ public class Product {
         this.name = name;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
